[feature] address modes

+++ include/m_memory.h
@@ -6,7 +6,7 @@
 #define M_MEMORY_SIZE (1 << 16)
 
 typedef struct m_memory {
-	uint16_t mem[M_MEMORY_SIZE];
+	uint8_t mem[M_MEMORY_SIZE];
 } m_memory_t;
 
 void m_reset(m_memory_t *mem);
